lexeme_entry disambig finished
