+++ requirements/base_requirements.txt
@@ -1,2 +1,3 @@
 Django==1.8
 unicode-slugify==0.1.3
+lazy==1.2
