+++ wordengine/global_const.py
@@ -51,6 +51,11 @@ REL_DIRECTION = (('F', 'Forward'),
 
 SPECIAL_CHARS = ('[', ']', '"', '|', '@')
 
+RELATION_TYPES = {
+    'pl': 'Plurale tantum',
+    'phr': 'Phrase'
+}
+
 # Common regexps
 
 RE_EXT_COMM = re.compile(r'(\*\d+)')
@@ -58,7 +63,13 @@ RE_PARAM = re.compile(r'(\[.*?\])')
 RE_COMMENT = re.compile(r'(\".*?\")')
 RE_REST_LIST = re.compile(r'\[.*\]')
 RE_REST_TUPLE = re.compile(r'\(.*\)')
-
+RE_DIALECT = re.compile(r'<(.*?)>')
+RE_FORM = re.compile(r'\{(.*?)\}')
+RE_COMMENT_NEW = re.compile(r'"""(.*?)"""')
+RE_GROUP_COMMENT = re.compile(r'^"""(.*?)"""')
+RE_SEM_GR = re.compile(r'^[\d\*]\.', re.M)
+RE_TRANSL = re.compile(r'^{(.*?)}', re.M)
+RE_EXAMPLE = re.compile(r'>>(.*?)', re.M)
 
 # Errors
 class WordcontrolException(Exception):
